+++ var/httpd/htdocs/js/Core.UI.ActionRow.js
@@ -176,7 +176,12 @@ Core.UI.ActionRow = (function (TargetNS) {
             TicketView = 'Small';
         }
 
-        $('#SelectAllTickets').bind('click', function () {
+        // Hide 'Select all' checkbox in Large and Medium view when table is empty.
+        if ((TicketView === 'Medium' || TicketView === 'Large') && $('#EmptyMessage' + TicketView).length === 1) {
+            $('#SelectAllTickets').closest('li').addClass('Hidden');
+        }
+
+        $('#SelectAllTickets').on('click', function () {
             var Status = $(this).prop('checked');
             $(TicketElementSelectors[TicketView]).prop('checked', Status).triggerHandler('click');
         });